fix(TFD-1241): Don't require keytab on worker nodes for fileio
The spec required the keytab to be installed on every worker node in the cluster, which is impractical for some clients.  Applying this patch causes the HDFS delegation token to be created on the driver for the principal/keytab and serialized to the nodes where it is reconstituted for all HDFS actions.

Since tokens are never reused with this system, it should also fix TCOMP-422